backend/requirements.txt
flask
flask-cors
boto3
pillow
torchvision
numpy
python-dotenv